requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2
PyPDF2==3.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/LLM
openai==1.7.2
anthropic==0.8.1

# Forecasting
statsmodels==0.14.1
prophet==1.1.5
scikit-learn==1.3.2

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6